+++ howso/direct/core.py
@@ -2231,21 +2231,6 @@ class HowsoCore:
             "session": session,
         })
 
-    def set_internal_parameters(self, trainee_id: str, params: Dict) -> None:
-        """
-        Sets specific model parameters in the Trainee.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        params : dict
-            A dictionary containing the internal parameters.
-        """
-        return self._execute(trainee_id, "set_internal_parameters", {
-            **params
-        })
-
     def set_feature_attributes(
         self,
         trainee_id: str,
@@ -2407,56 +2392,6 @@ class HowsoCore:
             "condition": condition,
             "num_cases": num_cases,
             "precision": precision,
-            "weight_feature": weight_feature,
-        })
-
-    def get_internal_parameters(
-        self,
-        trainee_id: str,
-        *,
-        action_feature: Optional[str] = None,
-        context_features: Optional[Iterable[str]] = None,
-        mode: Optional[Literal["robust", "full"]] = None,
-        weight_feature: Optional[str] = None,
-    ) -> Dict:
-        """
-        Get the parameters used by the Trainee.
-
-        If 'action_feature', 'context_features', 'mode', or 'weight_feature'
-        are specified, then the best hyperparameters analyzed in the Trainee
-        are the value of the 'hyperparameter_map' key, otherwise this value
-        will be the dictionary containing all the hyperparameter sets in
-        the Trainee.
-
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        action_feature : str, optional
-            If specified will return the best analyzed hyperparameters to
-            target this feature.
-        context_features : str, optional
-            If specified, will find and return the best analyzed hyperparameters
-            to use with these context features.
-        mode : str, optional
-            If specified, will find and return the best analyzed hyperparameters
-            that were computed in this mode.
-        weight_feature : str, optional
-            If specified, will find and return the best analyzed hyperparameters
-            that were analyzed using this weight feaure.
-
-        Returns
-        -------
-        dict
-            A dict including the either all of the Trainee's internal
-            parameters or only the best hyperparameters selected using the
-            passed parameters.
-        """
-        return self._execute(trainee_id, "get_internal_parameters", {
-            "action_feature": action_feature,
-            "context_features": context_features,
-            "mode": mode,
             "weight_feature": weight_feature,
         })
